+++ Movie-Trending-Week/Movie-Trending-Week/Models/TvShow.swift
@@ -14,7 +14,7 @@ struct TvShow: Codable {
     let posterPath: String
     let popularity: Double
     let firstAirDate: Date
-    let averageValueVote: Float
+    let averageValueVote: Double
     let voteTotal: Int
     
     enum CodingKeys: String, CodingKey {
